@@ -2,19 +2,20 @@ import 'package:flutter/cupertino.dart';
 import 'package:notes_app/src/blocs/notes_bloc.dart';
 
 class Provider extends InheritedWidget {
-  static Provider _instancia;
+  static Provider? _instancia;
   final NotesBloc _notesBloc = NotesBloc();
-  factory Provider({Key key, Widget child}) {
+  factory Provider({Key? key, Widget? child}) {
     if (_instancia == null) {
-      _instancia = Provider._internal(key: key, child: child);
+      _instancia = Provider._internal(key: key, child: child!);
     }
-    return _instancia;
+    return _instancia!;
   }
 
-  Provider._internal({Key key, Widget child}) : super(key: key, child: child);
+  Provider._internal({Key? key, required Widget child})
+      : super(key: key, child: child);
 
   static NotesBloc notesBloc(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<Provider>()._notesBloc;
+    return context.dependOnInheritedWidgetOfExactType<Provider>()!._notesBloc;
   }
 
   @override
